rpi3-edge-impulse-ppe-detection/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.4
botocore==1.39.4
certifi==2025.7.9
charset-normalizer==3.4.2
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.3.2
cycler==0.12.1
datasets==2.14.5
decorator==5.2.1
deprecation==2.1.0
dill==0.3.7
docopt==0.6.2
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2023.6.0
gdown==4.7.1
gitdb==4.0.12
GitPython==3.1.42
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
networkx==3.5
numpy==1.26.4
object-detection-metrics==0.4.post1
onnx==1.17.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
prettytable==3.11.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pydantic_yaml==1.4.0
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qai-hub==0.31.0
qai-hub-models==0.31.0
requests==2.32.4
requests-toolbelt==1.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
schema==0.7.5
scikit-image==0.25.2
scipy==1.16.0
semver==3.0.4
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.6.11
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
